/2022/movies -> /2022/uuid
