+++ src/routes/[year=year]/+page.ts
@@ -5,6 +5,6 @@ export const load: PageLoad = async ({ params, parent }) => {
   const { user } = await parent()
 
   if (user) {
-    throw redirect(302, `/${params.year}/movies`)
+    throw redirect(302, `/${params.year}/${user.uuid}`)
   }
 }
