--- src/routes/[year=year]/[uuid]/+page.ts
@@ -1,1 +1,0 @@
-export const ssr = true
