+++ src/routes/[year=year]/search/+page.svelte
@@ -56,7 +56,7 @@
 
   async function addRanking(movie: Movie) {
     await trpc.appendToRanking.query({ movieId: movie.id.toString(), year: parseInt(data.year) })
-    await goto(`/${data.year}/movies`)
+    await goto(`/${data.year}/${data.user?.uuid}`)
   }
 </script>
